Reduce decoder creation overheads for sun.nio.cs.ext Charsets
